+++ gulpfile.js/util/pattern-library/lib/renderLibraryPages.js
@@ -1,0 +1,15 @@
+var compileTemplate = require('./compileTemplate')
+var renderPagesFromTemplate = require('./renderPagesFromTemplate')
+var registerHandlebarsHelpers = require('./registerHandlebarsHelpers')
+
+var renderLibraryPages = function (config, files) {
+  registerHandlebarsHelpers(config.helpers)
+
+  return compileTemplate(config.template)
+    .then((compiledTemplate) =>
+      renderPagesFromTemplate(files, compiledTemplate, config.sourceBaseDir)
+    )
+    .catch((error) => error)
+}
+
+module.exports = renderLibraryPages