+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Changed
+
+- Replace the component library with the design pattern library [#822](https://github.com/hmrc/assets-frontend/pull/822)
+
 ## [2.252.0] - 2017-10-16
 
 ### Changed
@@ -14,7 +18,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Minor updates to input type="number" and fix for Firefox [#821](https://github.com/hmrc/assets-frontend/pull/821)
 - Updating `.notice-banner__content` styles so that they're actually visible [#820](https://github.com/hmrc/assets-frontend/pull/820)
 
-###Fixed
+### Fixed
 
 - Updated validatorFocus javascript to navigate to the input field in IE8 when there is a valudation error. [#805](https://github.com/hmrc/assets-frontend/pull/805)